+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `--force` option to `schedule:run` command.
+This option allow to run all tasks regardless of configured run time,
+part of issue [#11](https://github.com/lavary/crunz/issues/11) -
+PR [#120](https://github.com/lavary/crunz/pull/120) by [@PabloKowalczyk](https://github.com/PabloKowalczyk)
+
 ## 1.7.3 - 2018-06-15
 
 ### Fixed